:recycle: refactor: Passando as funcionalidades da pagina Logged para o componente Logged
