+++ src/components/Logged/styles.css
@@ -94,3 +94,26 @@
     color: rgb(242, 52, 52);
     font-size: 20px;
 }
+.button-update{
+    margin: 1em;
+    border: 1px solid silver;
+    border-radius: 5%;
+    padding: 0.6em 2.3em;
+    color: white;
+    background-color:  rgb(103, 161, 232);
+}
+.button-update:hover{
+    background-color:  rgb(93, 145, 208);
+}
+.popup-content label{
+    margin: 5px;
+}
+.popup-content input{
+    outline: none;
+    border-radius: 10px;
+    padding: 6px 25px;
+}
+.div-button-update{
+    display: flex;
+    justify-content: center;
+}
