--- src/pages/styles.css/popup.css
@@ -1,51 +1,0 @@
-.popup-wrapper{
-    background-color: rgba(0, 0, 0, .5);
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-}
-.none{
-    display: none;
-}
-.popup{
-    width: 100%;
-    max-width: 300px;
-    margin: 10% auto;
-    padding: 25px;
-    background: white;
-    position: relative;
-}
-.popup-close{
-    font-family: 'arial';
-    color: rgb(242, 52, 52);
-    font-size: 20px;
-    position: absolute;
-    top: 5px;
-    right: 10px;
-    cursor: pointer;
-}
-.button-update{
-    margin: 1em;
-    border: 1px solid silver;
-    border-radius: 5%;
-    padding: 0.6em 2.3em;
-    color: white;
-    background-color:  rgb(103, 161, 232);
-}
-.button-update:hover{
-    background-color:  rgb(93, 145, 208);
-}
-.popup-content label{
-    margin: 5px;
-}
-.popup-content input{
-    outline: none;
-    border-radius: 10px;
-    padding: 6px 25px;
-}
-.div-button-update{
-    display: flex;
-    justify-content: center;
-}
